Agentic-Chatbot-main_UI/requirements.txt
langchain==0.1.20
langgraph==0.0.47
langchain-core==0.1.52
langchain_community
langchain_groq
langchain_openai
faiss-cpu
streamlit
tavily-python
python-dotenv==1.0.1
llama-cpp-python==0.2.72
uvicorn[standard]==0.29.0
fastapi==0.110.1
loguru==0.7.2
pytest
websockets
sqlalchemy
psycopg2-binary